// frontend/src/types/enums.ts
export enum Role {
  USER = "USER",
  ADMIN = "ADMIN",
}

export enum Gender {
  MALE = "MALE",
  FEMALE = "FEMALE",
}

export enum RelationType {
  FATHER = "FATHER",
  MOTHER = "MOTHER",
  HUSBAND = "HUSBAND",
  WIFE = "WIFE",
  SON = "SON",
  DAUGHTER = "DAUGHTER",
  SON_OF_SON = "SON_OF_SON",
  DAUGHTER_OF_SON = "DAUGHTER_OF_SON",
  BROTHER = "BROTHER",
  SISTER = "SISTER",
  FULL_BROTHER = "FULL_BROTHER",
  FULL_SISTER = "FULL_SISTER",
  PATERNAL_BROTHER = "PATERNAL_BROTHER",
  PATERNAL_SISTER = "PATERNAL_SISTER",
  MATERNAL_BROTHER = "MATERNAL_BROTHER",
  MATERNAL_SISTER = "MATERNAL_SISTER",
  PATERNAL_GRANDFATHER = "PATERNAL_GRANDFATHER",
  PATERNAL_GRANDMOTHER = "PATERNAL_GRANDMOTHER",
  MATERNAL_GRANDFATHER = "MATERNAL_GRANDFATHER",
  MATERNAL_GRANDMOTHER = "MATERNAL_GRANDMOTHER",
  PATERNAL_UNCLE = "PATERNAL_UNCLE",
  PATERNAL_UNCLE_FULL = "PATERNAL_UNCLE_FULL",
  PATERNAL_UNCLE_PATERNAL = "PATERNAL_UNCLE_PATERNAL",
  MATERNAL_UNCLE = "MATERNAL_UNCLE",
  PATERNAL_AUNT = "PATERNAL_AUNT",
  MATERNAL_AUNT = "MATERNAL_AUNT",
  OTHER = "OTHER",
}

export enum CaseStatus {
  DRAFT = "DRAFT",
  CALCULATED = "CALCULATED",
  CLOSED = "CLOSED",
}

export enum BlockType {
  FULL_BLOCK = "FULL_BLOCK",
  PARTIAL_BLOCK = "PARTIAL_BLOCK",
}

export enum ReportType {
  PDF = "PDF",
  SUMMARY = "SUMMARY",
}

export enum Language {
  AR = "AR",
  EN = "EN",
}